refactor: simplify deploy package
